backend/requirements.txt
# Core
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# Redis
redis==5.2.1
hiredis==3.1.0

# WebSocket (Deriv)
websockets==13.1
aiohttp==3.11.10

# Data / Math
numpy==2.2.0
pandas==2.2.3

# Scheduling
apscheduler==3.10.4

# AI
groq==1.2.0

# Utilities
httpx==0.28.1
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
email-validator==2.2.0
python-multipart==0.0.19
loguru==0.7.3
tenacity==9.0.0
orjson==3.10.12
